refactor the tests.
move out 'test_schema from test_cli. Remove the 'test_cli/test_commands folder.
