+++ tests/test_cli/test_add/__init__.py
