+++ tests/test_cli/test_add/test_connection.py
