+++ tests/test_cli/test_delete.py
