+++ tests/test_cli/test_misc.py
