+++ tests/test_cli/test_remove/__init__.py
