+++ tests/test_cli/test_remove/test_protocol.py
