+++ tests/test_cli/test_run.py
@@ -35,7 +35,7 @@ from click.testing import CliRunner
 import aea.cli.common
 from aea.cli import cli
 from aea.configurations.base import DEFAULT_AEA_CONFIG_FILE, DEFAULT_CONNECTION_CONFIG_FILE
-from ...conftest import CLI_LOG_OPTION, CUR_PATH
+from tests.conftest import CLI_LOG_OPTION, CUR_PATH
 
 
 def test_run(pytestconfig):
