+++ tests/test_cli/test_scaffold/test_protocols.py
