+++ tests/test_configurations/test_schema.py
